move output files to x.go/common.go
